Completely recreated my project
Due to a situation where I lost access to the Components Tree in IntelliJ, I had to recreate the project so that I could have an easier time creating the forms for the project. I copied everything I had from the previous project that I could copy (.txt files, OOP classes, logic, etc) and moved everything to the new project.