Made a start on the hub page for the project
In the previous commit, I created all of the forms I will be needing for the project itself. What I have done in this commit is simple: I have made a start on creating the hub page and implemented a mouseClicked listener for the label. What I intend to do with this function is have the user click on the label to access the admin login screen (an idea I have ripped from my time working in retail).