Progress on StockDatabase
I have confirmed that data being read from .txt file into stock items array works as planned. I have used Breakpoints to verify that data is being read into the array exactly as I had designed it to be read.

My next step from here is to work on implementing the JList component of the StockDatabaseForm, making it so that it can display all of the items in the stock database.