Progress on Edit function!
Utilising an old way of doing things from SOFT152, I have implemented a function within the btnEdit method that will not only edit the array element linked to the selected row on tblStockData, but will also take the array post-edit and completely overwrite the testStockData.txt document using the contents of that array.

Next step: Work on refreshing the table so that it displays the new data after it has been edited.